+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -31,6 +31,7 @@ CHANGELOG
  * Deprecate the `framework.asset_mapper.provider` config option
  * Add `--exclude` option to the `cache:pool:clear` command
  * Add parameters deprecations to the output of `debug:container` command
+ * Change `framework.asset_mapper.importmap_polyfill` from a URL to the name of an item in the importmap
 
 6.3
 ---